Renaming find_board.py to hdmi2usb-mode-switch.py
Calling the script as atlys-XXXX will set --board=atlys
Calling the script as opsis-XXXX will set --board=opsis

Calling the script as XXXX-mode-switch will set good defaults for changing the
board mode. This functionality can require the tool to be run as root unless
the correct udev rules are installed.

Calling the script as XXXX-find-board will set good defaults for finding the
board on the system. This is useful for being being embedded in other scripts
and should never require root.